New Commit

+++ shopping-online/shopping-online/Context/Order_Details.cs
@@ -17,7 +17,7 @@ namespace shopping_online.Context
         public int Order_Details_id { get; set; }
         public Nullable<int> Order_id { get; set; }
         public Nullable<int> product_id { get; set; }
-        public Nullable<double> Order_Details_price { get; set; }
+        public Nullable<decimal> Order_Details_price { get; set; }
         public Nullable<int> Order_Details_num { get; set; }
         public Nullable<double> Order_Details_total_number { get; set; }
